feat(master): abstraction removed from can communication + parsing moved to manager callbacks
