+++ master/include/comm/manager.hpp
@@ -1,12 +1,7 @@
 #ifndef COMM_COMMS_HPP_
 #define COMM_COMMS_HPP_
 
-#ifdef EMBEDDED
-#include "comm/cancommunicator.hpp"
-#else
-#include "comm/vcancommunicator.hpp"
-#endif
-
+#include "comm/communicator.hpp"
 #include "logic/checkupManager.hpp"
 #include "logic/sensors.hpp"
 
@@ -24,16 +19,13 @@ class CommunicationManager {
     public:
     CommunicationManager(CheckupManager* checkupManager, Sensors* sensors);
 
-    void send_typical_message();
-
     void emergencySignalCallback(std::string msg);
     void missionFinishedCallback();
     void pcAliveCallback();
     void rlWheelCallback(double value);
     void hydraulicLineCallback(double value);
-    void r2dCallback();
-    void goCallback();
-    void bamocarTensionCallback(float value);
+    void resCallback(uint8_t *buf);
+    void bamocarCallback(uint8_t *buf);
     void steeringCallback();
 };
 
@@ -41,12 +33,7 @@ class CommunicationManager {
 
 CommunicationManager::CommunicationManager(CheckupManager* checkupManager, Sensors* sensors) 
     : checkupManager(checkupManager), sensors(sensors) {
-    #ifdef EMBEDDED /* These compiler flags allow the usage of one strategy or the
-        other depending on the environment being native (DESKTOP) or teensy (EMBEDDED) */
-    communicator = new CANCommunicator("can1");
-    #else
-    communicator = new VCANCommunicator("vcan1");
-    #endif
+    communicator = new Communicator("can1");
 };
 
 void CommunicationManager::emergencySignalCallback(std::string msg) {
@@ -69,16 +56,26 @@ void CommunicationManager::hydraulicLineCallback(double value) {
     sensors->updateHydraulic(value);
 }
 
-void CommunicationManager::r2dCallback() {
-    checkupManager->_ready2Drive = true;
+void CommunicationManager::resCallback(uint8_t *buf) {
+    bool emg_stop1 = buf[0] && 0x01;
+    bool emg_stop2 = buf[3] >> 7 && 0x01;
+    bool go_switch = (buf[0] >> 1) && 0x01;
+    bool go_button = (buf[0] >> 2) && 0x01;
+
+    if (go_button || go_switch)
+        checkupManager->_internalLogics.processGoSignal();
+    else if (emg_stop1 || emg_stop2)
+        checkupManager->_failureDetection.emergencySignal = true;
 }
 
-void CommunicationManager::goCallback(){
-    checkupManager->_internalLogics.processGoSignal();
-}
-
-void CommunicationManager::bamocarTensionCallback(float value) {
-    checkupManager->_failureDetection.bamocarTension = value;
+void CommunicationManager::bamocarCallback(uint8_t *buf) {
+    if (buf[0] == BTB_READY)
+        bool alive = true;
+        // TODO(andre): what to do with btb ready?
+    else if (buf[0] == VDC_BUS){
+        int battery_voltage = (buf[2] << 8) | buf[1];
+        checkupManager->_failureDetection.bamocarTension = battery_voltage;
+    }
 }
 
 void CommunicationManager::steeringCallback() {
